Add documentUpdater command
